fix: object/enact transaction state moved to motion detailed provider
